+++ KosmicTask/KosmicTask/source/FVFinderLabel.m
@@ -225,7 +225,7 @@ static void __FVLinearColorReleaseFunction(void *info)
 {
     // Apple preference for Finder label names
     NSDictionary *labelPrefs = [[NSUserDefaults standardUserDefaults] persistentDomainForName:@"com.apple.Labels"];
-    id name = [labelPrefs objectForKey:[NSString stringWithFormat:@"Label_Name_%d", label]];
+    id name = [labelPrefs objectForKey:[NSString stringWithFormat:@"Label_Name_%ld", (long)label]];
     // check the class, since this is private
     if ([name isKindOfClass:[NSString class]] == NO)
         name = nil;
@@ -234,7 +234,7 @@ static void __FVLinearColorReleaseFunction(void *info)
 
 + (NSString *)localizedNameForLabel:(NSInteger)label
 {
-    NSAssert1(label <= 7, @"Invalid Finder label %d (must be in the range 0--7)", label);
+    NSAssert1(label <= 7, @"Invalid Finder label %ld (must be in the range 0--7)", (long)label);
     NSString *name = [self _preferenceNameForLabel:label];
     NSBundle *bundle = [NSBundle bundleForClass:[FVFinderLabel self]];
     if (nil == name) {
@@ -375,7 +375,7 @@ static void ClipContextToCircleCappedPathInRect(CGContextRef context, CGRect rec
 
 + (void)drawFinderLabel:(NSUInteger)label inRect:(CGRect)rect ofContext:(CGContextRef)context flipped:(BOOL)isFlipped roundEnds:(BOOL)flag
 {
-    NSAssert1(label <= 7, @"Invalid Finder label %d (must be in the range 0--7)", label);
+    NSAssert1(label <= 7, @"Invalid Finder label %lu (must be in the range 0--7)", (unsigned long)label);
     
     CGLayerRef layerToDraw = NULL;
     
@@ -445,7 +445,7 @@ static void ClipContextToCircleCappedPathInRect(CGContextRef context, CGRect rec
 {
     FSRef fileRef;
     
-    NSAssert1(label <= 7, @"Invalid Finder label %d (must be in the range 0--7)", label);
+    NSAssert1(label <= 7, @"Invalid Finder label %lu (must be in the range 0--7)", (unsigned long)label);
         
     if ([aURL isFileURL] && CFURLGetFSRef((CFURLRef)aURL, &fileRef)) {
